chore(chromatic): restructure
