+++ .github/workflows/chromatic.yml
@@ -1,6 +1,5 @@
 name: Chromatic
 
-# Visual regression testing: builds the storybook workspace and sends it to Chromatic.
 on:
   # Manual full build: TurboSnap off, gate bypassed.
   workflow_dispatch:
@@ -10,7 +9,6 @@ on:
   pull_request:
     types: [opened, synchronize, reopened, ready_for_review]
 
-# Cancel an in-progress run when a newer push lands on the same ref; the latest wins.
 concurrency:
   group: chromatic-${{ github.ref }}
   cancel-in-progress: true
@@ -54,11 +52,8 @@ jobs:
             pnpm-lock.yaml
           files_ignore: |
             storybook/chromatic.config.json
-            storybook/.storybook/main.ts
 
-      # No UI changed (or the changesets version PR - version/CHANGELOG bumps
-      # only), so Chromatic is skipped. Post a passing "UI Tests" status
-      # ourselves so a required check still reports.
+      # Post the passing status ourselves, so a required check still reports.
       - name: No UI changes detected
         if:
           (steps.changed-files.outputs.any_changed != 'true' || github.head_ref
@@ -73,8 +68,7 @@ jobs:
             -f context='UI Tests' \
             -f description='No UI-affecting changes; Chromatic skipped'
 
-      # Installs dependencies and runs `pnpm build`. Stories import workspace
-      # packages by name, so those have to be built before Storybook is.
+      # Stories import workspace packages by name, so those build first.
       - name: Installing dependencies and building packages
         if: (steps.changed-files.outputs.any_changed == 'true' ||
           github.event_name == 'workflow_dispatch') && github.head_ref !=
@@ -89,11 +83,8 @@ jobs:
         with:
           # Per-surface name; the Cypress project's token lives in main.yml's test_playground.
           projectToken: ${{ secrets.CHROMATIC_TOKEN_UI_COMPONENTS }}
-          # Run inside the storybook workspace rather than pointing
-          # buildScriptName at a root wrapper script. Chromatic controls the
-          # build output location by appending `--output-dir`, and that argument
-          # is silently dropped through `pnpm run <root-script> -- <args>`, so a
-          # wrapper would upload from an empty directory.
+          # Not a root wrapper script: `pnpm run <script> -- <args>` swallows the
+          # `--output-dir` Chromatic appends, so it would upload an empty directory.
           workingDir: storybook
           storybookBaseDir: storybook
           buildScriptName: build
