+++ packages/application-components/package.json
@@ -73,15 +73,11 @@
   },
   "devDependencies": {
     "@apollo/client": "catalog:apollo",
-    "@commercetools-uikit/link": "catalog:repo",
-    "@commercetools-uikit/text-field": "catalog:repo",
-    "@commercetools-uikit/text-input": "catalog:repo",
     "@testing-library/dom": "catalog:test",
     "@testing-library/react": "catalog:test",
     "@types/jest": "catalog:test",
     "@types/raf-schd": "catalog:",
     "@types/testing-library__jest-dom": "catalog:test",
-    "formik": "catalog:react",
     "jest": "catalog:test",
     "msw": "catalog:test",
     "react": "catalog:react",
