+++ storybook/AGENTS.md
@@ -6,25 +6,36 @@ See root `AGENTS.md` for monorepo-wide context.
 
 Internal Storybook that hosts the visual regression tests Chromatic screenshots.
 
+**This is not a docs site.** The stories are VRT fixtures: `AllVariants` stacks, no
+args, no prop tables. Component documentation lives on
+[docs.commercetools.com](https://docs.commercetools.com/merchant-center-customizations).
+To browse the components, Chromatic hosts the built Storybook for every CI run and
+gives a permalink per branch.
+
 ## Key Context
 
-- **Stories are colocated in `packages/*`**, not here. This workspace holds only the Storybook config, decorators and shared helpers. The one exception is `application-icons`, which reads SVGs out of `packages/assets` and so has no owning component package.
+- **Stories live here**, in `src/stories/`, with shared scaffolding in `src/fixtures/`. They import components through public entrypoints (`@commercetools-frontend/*`), so the published packages carry no test scaffolding and nothing in them depends on this private workspace.
 - **Every story is captured.** Chromatic's capture-by-default stands, because every story here exists to be snapshotted. Opt one out with `chromatic: { disableSnapshot: true }`. There are no tags or parameters to set.
 - **Two decorators, both global.** `providers-decorator` supplies the stack these components assume (Apollo with a seeded cache, FlopFlip, react-intl, a `MemoryRouter`, plus `ThemeProvider` and `PortalsContainer`). `padding-decorator` adds 1rem, because Chromatic crops to rendered content and edge-painted focus rings would clip.
-- **`VisualSpec` and `VisualSpecGroup`** (`src/helpers/`) wrap each captured state, imported via the `@/storybook-helpers` alias. Picking the wrong one is the most common way a story renders wrong.
+- **`VisualSpec` and `VisualSpecGroup`** (`src/helpers/`) wrap each captured state. Picking the wrong one is the most common way a story renders wrong.
 - Chromatic runs against the **`app-kit-components`** project. The Cypress playground suite is a separate Chromatic project and does not involve this workspace.
 
 ## How To Work Here
 
-| Task          | Command                                                       |
-| ------------- | ------------------------------------------------------------- |
-| Dev server    | `pnpm storybook:start` — port 6006                            |
-| Build         | `pnpm storybook:build` — this is what Chromatic builds in CI  |
-| Run Chromatic | `pnpm --dir storybook exec chromatic --project-token=<token>` |
+| Task          | Command                                                      |
+| ------------- | ------------------------------------------------------------ |
+| Dev server    | `pnpm storybook:start` — port 6006                           |
+| Build         | `pnpm storybook:build` — this is what Chromatic builds in CI |
+| Browse hosted | TODO: paste the Chromatic permalink for `main`               |
+
+**Don't run Chromatic locally.** A local run publishes into the same project and lands
+in its build history, where it can take a baseline that no PR reviewed. Debug against
+`pnpm storybook:start` and let the PR build do the comparison.
 
 ### Adding a VRT story
 
-Colocate `<component>.stories.tsx` next to the component source in `packages/*`, then:
+Add `<component>.stories.tsx` to `src/stories/`, importing the component from its
+public entrypoint (`@commercetools-frontend/*`), never a deep path. Then:
 
 1. `title: 'Application Components/<Name>'`, and `component:` when the export is a single component.
 2. Wrap each state in the right helper:
@@ -45,7 +56,7 @@ comes out **0px wide and captures nothing**.
 
 ## Gotchas
 
-- Stories in `packages/*` import Storybook types through a root-`tsconfig.json` `paths` entry, because pnpm keeps `@storybook/react-vite` in this workspace. Type-only imports, so nothing resolves it at runtime.
+- `src/stubs/supported-locales.ts` exists because `packages/l10n` reads its locale list through `babel-plugin-preval`, which Vite doesn't run, so the import lands on a CJS file with no default export and the `application-shell` barrel fails to load.
 - `src/stubs/msw.ts` exists because Storybook 9.1 injects a module mocker that imports `msw/browser`, which msw v1 doesn't have. Delete it when msw reaches v2.
 - `globals.css` must import `custom-properties.css`. `resets.css` sets `font-family: var(--font-family)` with no inline fallback, so without it everything renders in the browser's default serif.
 - app-kit's own tokens (`--margin-for-page-content` and friends) are defined by the `ThemeProvider` in `providers-decorator`, and have **no** inline fallbacks. Remove it and page layouts lose their padding silently.
