+++ storybook/package.json
@@ -5,16 +5,26 @@
   "private": true,
   "scripts": {
     "start": "storybook dev -p 6006",
-    "build": "NODE_ENV=production storybook build",
-    "chromatic": "npx chromatic --project-token=chpt_7144b9fd8ba78b8 --dry-run"
+    "build": "NODE_ENV=production storybook build"
   },
   "dependencies": {
     "@apollo/client": "catalog:apollo",
     "@commercetools-frontend/application-components": "workspace:^",
+    "@commercetools-frontend/application-shell": "workspace:^",
+    "@commercetools-frontend/assets": "workspace:^",
     "@commercetools-frontend/constants": "workspace:^",
+    "@commercetools-frontend/react-notifications": "workspace:^",
+    "@commercetools-uikit/card": "catalog:repo",
     "@commercetools-uikit/design-system": "catalog:repo",
     "@commercetools-uikit/grid": "catalog:repo",
+    "@commercetools-uikit/icon-button": "catalog:repo",
     "@commercetools-uikit/icons": "catalog:repo",
+    "@commercetools-uikit/link": "catalog:repo",
+    "@commercetools-uikit/spacings": "catalog:repo",
+    "@commercetools-uikit/text": "catalog:repo",
+    "@commercetools-uikit/text-field": "catalog:repo",
+    "@commercetools-uikit/text-input": "catalog:repo",
+    "@commercetools/nimbus": "catalog:nimbus",
     "@emotion/babel-plugin": "catalog:build",
     "@emotion/react": "catalog:build",
     "@emotion/styled": "catalog:build",
@@ -27,6 +37,7 @@
     "@types/react-router-dom": "catalog:react",
     "@vitejs/plugin-react": "catalog:build",
     "chromatic": "catalog:",
+    "formik": "catalog:react",
     "graphql": "catalog:apollo",
     "react": "catalog:react",
     "react-dom": "catalog:react",
