+++ storybook/src/helpers/lorem.ts
@@ -1,5 +1,3 @@
-// Shared by every "long title and subtitle" state, across all the page families.
-
 export const longTitle =
   'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.';
