+++ storybook/src/stories/confirmation-dialog.stories.tsx
@@ -1,8 +1,8 @@
-import { VisualSpecGroup } from '@/storybook-helpers';
 import type { Meta, StoryObj } from '@storybook/react-vite';
+import { ConfirmationDialog } from '@commercetools-frontend/application-components';
 import Spacings from '@commercetools-uikit/spacings';
 import Text from '@commercetools-uikit/text';
-import ConfirmationDialog from './confirmation-dialog';
+import { VisualSpecGroup } from '../helpers';
 
 type TExampleProps = {
   portalId: string;
