+++ storybook/src/stories/page-not-found.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import PageNotFound from './page-not-found';
+import { PageNotFound } from '@commercetools-frontend/application-components';
 
 const meta: Meta<typeof PageNotFound> = {
   title: 'Application Components/PageNotFound',
