+++ storybook/src/stories/tabular-modal-page.stories.tsx
@@ -1,6 +1,6 @@
 import type { ComponentProps } from 'react';
-import { CUSTOM_VIEW_LOCATORS } from '@/storybook-helpers';
 import type { Meta, StoryObj } from '@storybook/react-vite';
+import { TabularModalPage } from '@commercetools-frontend/application-components';
 import IconButton from '@commercetools-uikit/icon-button';
 import {
   BinLinearIcon,
@@ -11,8 +11,8 @@ import {
   TabularPageContent,
   TabularPageCustomTitleRow,
   TabularPageTabControls,
-} from '../../tabular-page-fixtures';
-import TabularModalPage from './tabular-modal-page';
+} from '../fixtures/tabular-page';
+import { CUSTOM_VIEW_LOCATORS } from '../helpers';
 
 const routePath = '/tabular-modal-page';
